+++ db/dumps/current_schema.sql
@@ -3,7 +3,7 @@ begin;
 -- PostgreSQL database dump
 --
 
-\restrict OqeeQ5eK7zpD1amgL0XSdmnlgh7NJVpOMOXAk7bXfXssJwmxeRjddco2LBjq3ur
+\restrict PBvrJMoqOp4IV2J4bNOrfaMaGPMUf1aAGcO7PgEc26CIQw7Vxn7ufPfFLKdgDzw
 
 -- Dumped from database version 15.14
 -- Dumped by pg_dump version 15.14
@@ -13575,6 +13575,6 @@ ALTER TABLE ONLY ref_data.redcap_data_dictionary_history
 -- PostgreSQL database dump complete
 --
 
-\unrestrict OqeeQ5eK7zpD1amgL0XSdmnlgh7NJVpOMOXAk7bXfXssJwmxeRjddco2LBjq3ur
+\unrestrict PBvrJMoqOp4IV2J4bNOrfaMaGPMUf1aAGcO7PgEc26CIQw7Vxn7ufPfFLKdgDzw
 
 commit;